# docs/swagger.yaml
definitions:
  dto.AddPlatformDTO:
    properties:
      name:
        items:
          type: string
        type: array
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: this is about novel swagger API
  title: Cabbage server API
  version: "1.0"
paths:
  /v1/api/admin/profile/platform/add:
    post:
      consumes:
      - application/json
      description: 添加新的社交平台
      parameters:
      - description: platforms
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddPlatformDTO'
      responses: {}
      summary: create new social platform
      tags:
      - admin
...
  /v1/api/admin/stat/all:
    get:
      consumes:
      - application/json
      description: 统计平台信息 今日新增贴子数 评论数 用户数  当月新增用户数 评论数 用户数
      parameters:
      - enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        in: query
        name: month
        required: true
        type: integer
      responses: {}
      summary: get stat data
      tags:
      - admin
...
swagger: "2.0"